Split `check` for `javac` and `scalac` out into their own modules.
[ci skip-rust]
[ci skip-build-wheels]
