+++ src/python/pants/backend/java/compile/javac.py
@@ -8,10 +8,8 @@ from dataclasses import dataclass
 from itertools import chain
 
 from pants.backend.java.target_types import JavaSourceField
-from pants.core.goals.check import CheckRequest, CheckResult, CheckResults
 from pants.core.util_rules.archive import ZipBinary
 from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
-from pants.engine.addresses import Addresses
 from pants.engine.fs import (
     EMPTY_DIGEST,
     AddPrefix,
@@ -23,8 +21,7 @@ from pants.engine.fs import (
 )
 from pants.engine.process import BashBinary, FallibleProcessResult, Process, ProcessResult
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
-from pants.engine.target import CoarsenedTarget, CoarsenedTargets, FieldSet, SourcesField, Targets
-from pants.engine.unions import UnionRule
+from pants.engine.target import CoarsenedTarget, CoarsenedTargets, FieldSet, SourcesField
 from pants.jvm.compile import CompiledClassfiles, CompileResult, FallibleCompiledClassfiles
 from pants.jvm.compile import rules as jvm_compile_rules
 from pants.jvm.jdk_rules import JdkSetup
@@ -48,10 +45,6 @@ class JavacFieldSet(FieldSet):
     required_fields = (JavaSourceField,)
 
     sources: JavaSourceField
-
-
-class JavacCheckRequest(CheckRequest):
-    field_set_type = JavacFieldSet
 
 
 @dataclass(frozen=True)
@@ -249,32 +242,8 @@ async def compile_java_source(
     )
 
 
-@rule(desc="Check javac compilation", level=LogLevel.DEBUG)
-async def javac_check(request: JavacCheckRequest) -> CheckResults:
-    coarsened_targets = await Get(
-        CoarsenedTargets, Addresses(field_set.address for field_set in request.field_sets)
-    )
-
-    resolves = await MultiGet(
-        Get(CoursierResolveKey, Targets(t.members)) for t in coarsened_targets
-    )
-
-    results = await MultiGet(
-        Get(
-            FallibleCompiledClassfiles,
-            CompileJavaSourceRequest(component=target, resolve=resolve),
-        )
-        for target, resolve in zip(coarsened_targets, resolves)
-    )
-
-    # NB: We don't pass stdout/stderr as it will have already been rendered as streaming.
-    exit_code = next((result.exit_code for result in results if result.exit_code != 0), 0)
-    return CheckResults([CheckResult(exit_code, "", "")], checker_name="javac")
-
-
 def rules():
     return [
         *collect_rules(),
         *jvm_compile_rules(),
-        UnionRule(CheckRequest, JavacCheckRequest),
     ]
