+++ src/python/pants/backend/scala/goals/BUILD
@@ -1,0 +1,1 @@
+python_sources()
